+++ app/src/components/ContentContainer.vue
@@ -9,4 +9,11 @@
     background-color: var(--color-secondary);
     padding: 2rem;
   }
+
+  @media(min-width: 992px) {
+    .container {
+      margin: 0 auto;
+      max-width: 960px;
+    }
+  }
 </style>