+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.2] - 2024-02-08
+### Updated
+- Updates to readme file
+- Content updates
+- Minor style updates
+
+### Added
+- CardsComponent to display sets of card data
+  - still needs better styling
+- Unit tests for card components
+
 ## [1.0.1] - 2024-02-08
 ### Added
 - Added google analytics